Correct plot guess issue with simultaneous fits Re #29043
This corrects an issue whereby plot guess was causing errors for simultaneous fits by expecting one workspace and instead receiving a list. This corrects this by using just the first workspace in the list if a list is given. Whilst not a perfect solution this is ok as you only use the given workspace as an x axis for the evaluation so you just need a workspace not all of them or a specific one.