Fix issue on mirroring in IndirectQENS for IN16B
When using the mirroring option (except 0-2-3) when reducing IN16B
QENS indirect data, coupled with the CropDeadMonitorChannels
option set to True, there is an underlying hypothesis that the two
 wings are symetrical. It turns out it can not be the case, and
thus the channels are cropped unevenly, resulting in a crash.

Now the channels are cropped so they are equal and no data is lost.