Fix up various failing system tests
Fixes up various failing system tests which were highlighting some
obscure bugs and edge cases.

Skip SystemTests for Single Reduction Version 2 since this isn't used as
it produces slight differences in some obscure modes. It also
fundamentally works differently to Version 1 so requires more planning
as we can't just swap the loops around.

Since the SANS group is not using it we can come back to it when there
is such demand as per agile principles